Выносим класс из main
