+++ src/main.js
@@ -11,53 +11,12 @@ import MovieDetailsPopupView from './view/movie-detail-popup';
 import {createMovieMock} from './mock/movie-mock';
 import {createUserProfileMock} from './mock/user-profile-mock';
 import {createFilters} from './mock/filters-mock';
-import {render, RenderPosition} from "./utils.js";
+import {render, RenderPosition} from './utils.js';
+import ArrayChunkIterator from './array-chunk-iterator';
 
 const ALL_MOVIES_BOARD_CARDS_PORTION_COUNT = 5;
 const EXTRA_BOARDS_MOVIES_CARDS_COUNT = 2;
 const MOVIES_COUNT = 20;
-
-/**
- * Итератор, выдающий элементы исходного массива группами, заданной размерности
- */
-class ArrayChunkIterator {
-  /**
-   * @param {*} arr исходный массив
-   * @param {number} chunkSize максимальная размерность массива, выдаваемого за одну итерацию
-   */
-  constructor(arr = [], chunkSize = 1) {
-    this._arr = arr;
-    this._chunkSize = chunkSize;
-    this._counter = 0;
-    this._isDone = arr.length === 0;
-  }
-
-  /**
-   * Возвращает группу элементов исходного массива
-   * @return {array}
-   */
-  next() {
-    if (this._isDone) {
-      throw new RangeError(`Out of range`);
-    }
-
-    const arr = this._arr.slice(this._counter, this._counter + this._chunkSize);
-    this._counter += arr.length;
-
-    if (this._counter >= this._arr.length) {
-      this._isDone = true;
-    }
-
-    return arr;
-  }
-
-  /**
-   * Признак того, что все элементы массива проитерированы
-   */
-  get isDone() {
-    return this._isDone;
-  }
-}
 
 const movies = Array(MOVIES_COUNT).fill().map(createMovieMock);
 const userProfile = createUserProfileMock();
